requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
google-generativeai==0.8.3
uvicorn==0.31.0